+++ www/fns/Users/delete.php
@@ -3,8 +3,17 @@
 namespace Users;
 
 function delete ($mysqli, $id_users) {
-    $mysqli->query("delete from users where id_users = $id_users");
-    $dirname = __DIR__."/../../users/$id_users";
-    rmdir("$dirname/files");
-    rmdir($dirname);
+
+    $sql = "delete from users where id_users = $id_users";
+    $mysqli->query($sql) || trigger_error($mysqli->error);
+
+    $rmdir = function ($dirname) {
+        if (is_dir($dirname)) rmdir($dirname);
+    };
+
+    $userDir = __DIR__."/../../users/$id_users";
+    $rmdir("$userDir/files");
+    $rmdir("$userDir/received-files");
+    $rmdir($userDir);
+
 }
